requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
asyncpg>=0.30.0
pydantic>=2.9.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.1
fastapi-mail>=1.4.1
email-validator>=2.1.0
